[FrameworkBundle][Serializer] Add SerializerBundle
