+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/cache_pools_without_sections.php
@@ -2,7 +2,6 @@
 
 $container->loadFromExtension('framework', [
     'validation' => false,
-    'serializer' => false,
     'property_info' => false,
     'messenger' => false,
     'scheduler' => false,
