+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/full.php
@@ -47,21 +47,6 @@ $container->loadFromExtension('framework', [
         'enabled' => true,
         'email_validation_mode' => 'html5',
     ],
-    'serializer' => [
-        'enabled' => true,
-        'enable_attributes' => true,
-        'name_converter' => 'serializer.name_converter.camel_case_to_snake_case',
-        'circular_reference_handler' => 'my.circular.reference.handler',
-        'max_depth_handler' => 'my.max.depth.handler',
-        'default_context' => ['enable_max_depth' => true],
-        'named_serializers' => [
-            'api' => [
-                'include_built_in_normalizers' => true,
-                'include_built_in_encoders' => true,
-                'default_context' => ['enable_max_depth' => false],
-            ],
-        ],
-    ],
     'property_info' => true,
     'request' => [
         'formats' => [
