+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/profiler.php
@@ -4,7 +4,4 @@ $container->loadFromExtension('framework', [
     'profiler' => [
         'enabled' => true,
     ],
-    'serializer' => [
-        'enabled' => true,
-    ],
 ]);
