--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/serializer_without_translator.php
@@ -1,10 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'serializer' => [
-        'enabled' => true,
-    ],
-    'translator' => [
-        'enabled' => false,
-    ],
-]);
