+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/cache_pools_without_sections.yml
@@ -1,6 +1,5 @@
 framework:
     validation: false
-    serializer: false
     property_info: false
     messenger: false
     scheduler: false
