+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/profiler.yml
@@ -1,5 +1,3 @@
 framework:
     profiler:
         enabled: true
-    serializer:
-        enabled: true
