--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/serializer_enabled.yml
@@ -1,3 +1,0 @@
-framework:
-    serializer:
-        enabled: true
