--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/serializer_without_translator.yml
@@ -1,5 +1,0 @@
-framework:
-    serializer:
-        enabled: true
-    translator:
-        enabled: false
