+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -71,7 +71,7 @@
         "symfony/scheduler": "^8.2",
         "symfony/security-bundle": "^7.4|^8.0",
         "symfony/semaphore": "^8.2",
-        "symfony/serializer": "^7.4|^8.0",
+        "symfony/serializer": "^8.2",
         "symfony/stopwatch": "^7.4|^8.0",
         "symfony/string": "^7.4|^8.0",
         "symfony/translation": "^7.4|^8.0",
@@ -101,7 +101,7 @@
         "symfony/notifier": "<8.2",
         "symfony/security-csrf": "<7.4",
         "symfony/semaphore": "<8.2",
-        "symfony/serializer": "<7.4",
+        "symfony/serializer": "<8.2",
         "symfony/messenger": "<8.2",
         "symfony/mime": "<8.2",
         "symfony/object-mapper": "<8.2",
