+++ src/Symfony/Component/Serializer/Resources/config/serializer_debug.php
