+++ src/Symfony/Component/Serializer/composer.json
@@ -27,9 +27,10 @@
         "symfony/cache": "^7.4|^8.0",
         "symfony/config": "^7.4|^8.0",
         "symfony/console": "^7.4|^8.0",
-        "symfony/dependency-injection": "^7.4|^8.0",
+        "symfony/dependency-injection": "^8.2",
         "symfony/error-handler": "^7.4|^8.0",
         "symfony/filesystem": "^7.4|^8.0",
+        "symfony/finder": "^7.4|^8.0",
         "symfony/form": "^7.4|^8.0",
         "symfony/http-foundation": "^7.4|^8.0",
         "symfony/http-kernel": "^7.4|^8.0",
@@ -48,6 +49,7 @@
     "conflict": {
         "phpdocumentor/reflection-docblock": "<5.2|>=7",
         "phpdocumentor/type-resolver": "<1.5.1",
+        "symfony/dependency-injection": "<8.2",
         "symfony/property-access": "<8.1",
         "symfony/property-info": "<7.4.15",
         "symfony/type-info": "<7.4"
